Introduce --reference-only flag
The --reference-only flag allows to avoid any network access at an
initial repo sync operation. I.e. it enforces that a local mirror url
of the repo is used instead of the (potential network) url as given in
the manifest.
The local mirror url is composed of --reference directory
plus <project name as of the manifest>.git.

Additinally, some logger.debug outputs are added.
Additionally, a consistency check is done that --reference-only flag set
requires --reference given as well.

The new attribute reference_only is stored as Flag in the manifest repo
CONFIG file.